feat: rest infra changes
