+++ conf/rest/9.13.0/disk.yaml
@@ -24,14 +24,14 @@ counters:
   - ^node.name                => node
   - ^home_node.name           => owner_node
 
+endpoints:
+  - query: api/private/cli/disk
+    counters:
+      - ^^uid => uid
+      - sectors_read => stats_sectors_read
+      - sectors_written => stats_sectors_written
+
 plugins:
-  - RestGaps:
-      - disk:
-          query: api/private/cli/disk
-          counters:
-            - ^^uid => uid
-            - sectors_read => stats_sectors_read
-            - sectors_written => stats_sectors_written
   - LabelAgent:
       value_to_num:
         - new_status outage - - `0` #ok_value is empty value, '-' would be converted to blank while processing.
