+++ conf/rest/9.13.0/volume.yaml
@@ -2,10 +2,6 @@ name:                     Volume
 query:                    api/storage/volumes
 object:                   volume
 
-fields:
-  - space
-  - autosize
-
 counters:
   - autosize.maximum => autosize_maximum_size
   - autosize.grow_threshold => autosize_grow_threshold_percent
@@ -36,25 +32,27 @@ counters:
   - space.snapshot.space_used_percent    => snapshot_reserve_used_percent
   - ^snapshot_policy.name  => snapshot_policy
   - ^state   => state
+  - hidden_fields:
+    - space
+    - autosize
 
+endpoints:
+  - query: api/private/cli/volume
+    counters:
+      - ^^instance_uuid => instance_uuid
+      - ^node  => node
+      - ^is_sis_volume => is_sis_volume
+      - snapshot_count => snapshot_count
+      - compression_space_saved_percent => sis_compress_saved_percent
+      - dedupe_space_saved_percent => sis_dedup_saved_percent
+      - compression_space_saved => sis_compress_saved
+      - dedupe_space_saved => sis_dedup_saved
+      - sis_space_saved => sis_total_saved
+      - sis_space_saved_percent => sis_total_saved_percent
+      - files => inode_files_total
+      - files_used => inode_files_used
 
 plugins:
-  - RestGaps:
-      - volume:
-          query: api/private/cli/volume
-          counters:
-            - ^^instance_uuid => instance_uuid
-            - ^node  => node
-            - ^is_sis_volume => is_sis_volume
-            - snapshot_count => snapshot_count
-            - compression_space_saved_percent => sis_compress_saved_percent
-            - dedupe_space_saved_percent => sis_dedup_saved_percent
-            - compression_space_saved => sis_compress_saved
-            - dedupe_space_saved => sis_dedup_saved
-            - sis_space_saved => sis_total_saved
-            - sis_space_saved_percent => sis_total_saved_percent
-            - files => inode_files_total
-            - files_used => inode_files_used
   - LabelAgent:
       value_to_num:
         - new_status state online online `0`
